+++ ev-store-be/src/main/java/ge/evstore/ev_store/repository/CategoryRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     @Query("SELECT c FROM Category c WHERE c.parentCategory = :id")
-    List<Category> findByParentCategory(Integer id);
+    List<Category> findByParentCategory(Long id);
 }
